Move the two extra shell protocols out of WaylandState
Popups and the session lock are the surfaces that are not the layer shell,
and each carries a rule the rest of the platform layer does not have to know
about. They become `platform::popups` and `platform::lock`.

A popup does not choose where it goes: the client states an anchor, a
gravity and how it may be adjusted, and the compositor answers with a
configure — which is why growing one means asking to be repositioned rather
than resizing, and why teardown is ordered deepest-first (destroying a
parent before its child is a protocol error). Locking is a grant, not a
local success: the compositor may refuse, so the outcome arrives as an event
the main loop drains instead of as a return value, and `LockEvent` moves
next to the code that produces it.

wayland.rs 1407 -> 897 lines, 20 -> 17 fields.
